backend/requirements.txt
fastapi
uvicorn
openpyxl
docxtpl
python-docx
pydantic[email]
sqlalchemy
psycopg2-binary
alembic
python-dotenv
